pre-campaign-intelligence/pyproject.toml
[project]
name = "pre-campaign-intelligence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.136.1",
    "google-genai>=1.75.0",
    "numpy>=2.4.4",
    "pydantic>=2.13.3",
    "scikit-learn>=1.8.0",
    "vadersentiment>=3.3.2",
]
